+++ nix/hevm-tests/smt-checker.nix
@@ -95,6 +95,7 @@ let
     "loops/while_1_infinite.sol"
 
     # --- hevm timeout ---
+
     "types/array_dynamic_3_fail.sol"
     "types/array_aliasing_storage_1.sol"
     "types/array_aliasing_memory_1.sol"
@@ -189,6 +190,11 @@ let
     "typecast/cast_smaller_3.sol"
     "types/mapping_as_parameter_1.sol"
     "types/mapping_as_local_var_1.sol"
+
+    # --- solc 0.8 ---
+
+    "modifiers/modifier_code_after_placeholder.sol" # assertion can't trigger post 0.8 due to overflow checking
+    "special/many.sol" # no longer compiles
 
   ];