+++ overlay.nix
@@ -71,7 +71,8 @@ in rec {
           //
           fetchSolcVersions { owner = "dapphub"; attr = "unreleased_" + super.system; }
         );
-  solc = self.pkgs.runCommand "solc" { } "mkdir -p $out/bin; ln -s ${solc-static-versions.solc_0_6_7}/bin/solc-0.6.7 $out/bin/solc";
+
+  solc = self.pkgs.runCommand "solc" { } "mkdir -p $out/bin; ln -s ${solc-static-versions.solc_0_8_6}/bin/solc-0.8.6 $out/bin/solc";
 
   solc-static-versions =
     let